chore: move file.py to utils file
